+++ EssentialFeed/EssentialFeedTests/Feed Cache/Helpers/FeedImageDataStoreSpy.swift
@@ -1,0 +1,39 @@
+//
+//  FeedImageDataStoreSpy.swift
+//  EssentialFeedTests
+//
+//  Created by Alex Tapia on 24/04/22.
+//
+
+import Foundation
+import EssentialFeed
+
+class FeedImageDataStoreSpy: FeedImageDataStore {
+  enum Message: Equatable {
+    case insert(data: Data, for: URL)
+    case retrieve(dataFor: URL)
+  }
+  
+  private(set) var receivedMessages = [Message]()
+  private var retrievalCompletions = [(FeedImageDataStore.RetrievalResult) -> Void]()
+  
+  func insert(_ data: Data,
+              for url: URL,
+              completion: @escaping (FeedImageDataStore.InsertionResult) -> Void) {
+    receivedMessages.append(.insert(data: data, for: url))
+  }
+  
+  func retrieve(dataForURL url: URL,
+                completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
+    receivedMessages.append(.retrieve(dataFor: url))
+    retrievalCompletions.append(completion)
+  }
+  
+  func completeRetrieval(with error: Error, at index: Int = 0) {
+    retrievalCompletions[index](.failure(error))
+  }
+  
+  func completeRetrieval(with data: Data?, at index: Int = 0) {
+    retrievalCompletions[index](.success(data))
+  }
+}
